send notifcation when updating db
