+++ backend/updater/models.py
@@ -1,10 +1,6 @@
 from django.db import models
 from trips.models import Trip
 
-from django.db.models.signals import post_save
-from channels.layers import get_channel_layer
-
-
 class VehicleUpdate(models.Model):
     lat = models.FloatField()
     long = models.FloatField()
@@ -19,7 +15,6 @@ CAUSE = ((1, 'UNKNOWN_CAUSE'), (2, 'OTHER_CAUSE'),
          (7, 'HOLIDAY'), (8, 'WEATHER'), 
          (9, 'MAINTENANCE'), (10, 'CONSTRUCTION'), 
          (11, 'POLICE_ACTIVITY'), (12, 'MEDICAL_EMERGENCY'))
-
 EFFECT = ((1, 'NO_SERVICE'), (2, 'REDUCED_SERVICE'), (3, 'SIGNIFICANT_DELAYS'), 
           (4, 'DETOUR'), (5, 'ADDITIONAL_SERVICE'), (6, 'MODIFIED_SERVICE'), 
           (7, 'OTHER_EFFECT'), (8, 'UNKNOWN_EFFECT'), (9, 'STOP_MOVED'))
@@ -30,25 +25,3 @@ class Alert(models.Model):
     effect = models.IntegerField(choices=EFFECT)
     duration = models.DurationField()
     start_timestamp = models.TimeField()
-
-
-def save_alert_profile(sender, instance, **kwargs):
-    channel_layer = get_channel_layer()
-    print('\n\nhi\n\n')
-    print(f'{instance.affected_entity}_{instance.entity_id}')
-    channel_layer.send(f'{instance.affected_entity}_{instance.entity_id}', {
-        'type': 'notify_update',
-        'message': instance
-    })
-
-def save_vehicle_update_profile(sender, instance, **kwargs):
-    channel_layer = get_channel_layer()
-    channel_layer.send(f'trip_{instance.trip.id}', {
-        'type': 'notify_update',
-        'message': instance    
-    })
-
-
-post_save.connect(save_alert_profile, sender=Alert)
-post_save.connect(save_vehicle_update_profile, sender=VehicleUpdate)
-
